Task funiona de nuevo.

+++ src/main/java/com/todolist/controllers/GroupController.java
@@ -8,6 +8,7 @@ import com.todolist.dtos.ShowUser;
 import com.todolist.entity.Group;
 import com.todolist.entity.Task;
 import com.todolist.entity.User;
+import com.todolist.exceptions.BadRequestException;
 import com.todolist.filters.DateFilter;
 import com.todolist.filters.NumberFilter;
 import com.todolist.services.GroupService;
@@ -52,8 +53,14 @@ public class GroupController {
                                                   @RequestParam(required = false) NumberFilter numTasks,
                                                   @RequestParam(required = false) DateFilter createdDate) {
         String propertyOrder = order.charAt(0) == '+' || order.charAt(0) == '-' ? order.substring(1) : order;
-        Preconditions.checkArgument(Arrays.stream(ShowGroup.ALL_ATTRIBUTES.split(",")).anyMatch(prop -> prop.equalsIgnoreCase(propertyOrder)), "The order is invalid.");
-        Preconditions.checkArgument(Arrays.stream(fieldsGroup.split(",")).allMatch(field -> ShowGroup.ALL_ATTRIBUTES.toLowerCase().contains(field.toLowerCase())), "The fields are invalid.");
+        if (Arrays.stream(ShowGroup.ALL_ATTRIBUTES.split(",")).noneMatch(prop -> prop.equalsIgnoreCase(propertyOrder)))
+            throw  new BadRequestException("The order is invalid.");
+        if (!Arrays.stream(fieldsGroup.split(",")).allMatch(field -> ShowGroup.ALL_ATTRIBUTES.toLowerCase().contains(field.toLowerCase())))
+            throw new BadRequestException("The groups' fields are invalid.");
+        if (!Arrays.stream(fieldsUser.split(",")).allMatch(field -> ShowUser.ALL_ATTRIBUTES.toLowerCase().contains(field.toLowerCase())))
+            throw new BadRequestException("The users' fields are invalid.");
+        if (!Arrays.stream(fieldsTask.split(",")).allMatch(field -> ShowTask.ALL_ATTRIBUTES.toLowerCase().contains(field.toLowerCase())))
+            throw new BadRequestException("The tasks' fields are invalid.");
         List<ShowGroup> result = Lists.newArrayList(),
                 groups = groupService.findAllShowGroups(Sort.by(order.charAt(0) == '-' ? Sort.Direction.DESC : Sort.Direction.ASC, propertyOrder));
         if (limit == -1) limit = groups.size() - 1;
@@ -78,16 +85,20 @@ public class GroupController {
                                         @RequestParam(defaultValue = ShowUser.ALL_ATTRIBUTES) String fieldsUser,
                                         @RequestParam(defaultValue = ShowTask.ALL_ATTRIBUTES) String fieldsTask) {
         Group group = groupService.findGroupById(idGroup);
-        Preconditions.checkNotNull(group, "The group with idGroup " + idGroup + " does not exist.");
-        Preconditions.checkArgument(Arrays.stream(fieldsGroup.split(",")).allMatch(field -> ShowGroup.ALL_ATTRIBUTES.toLowerCase().contains(field.toLowerCase())), "The groups' fields are invalid.");
-        Preconditions.checkArgument(Arrays.stream(fieldsUser.split(",")).allMatch(field -> ShowUser.ALL_ATTRIBUTES.toLowerCase().contains(field.toLowerCase())), "The users' fields are invalid.");
-        Preconditions.checkArgument(Arrays.stream(fieldsTask.split(",")).allMatch(field -> ShowTask.ALL_ATTRIBUTES.toLowerCase().contains(field.toLowerCase())), "The tasks' fields are invalid.");
+        if (group == null) throw new BadRequestException("The group with idGroup " + idGroup + " does not exist.");
+        if (!Arrays.stream(fieldsGroup.split(",")).allMatch(field -> ShowGroup.ALL_ATTRIBUTES.toLowerCase().contains(field.toLowerCase())))
+            throw new BadRequestException("The groups' fields are invalid.");
+        if (!Arrays.stream(fieldsUser.split(",")).allMatch(field -> ShowUser.ALL_ATTRIBUTES.toLowerCase().contains(field.toLowerCase())))
+            throw new BadRequestException("The users' fields are invalid.");
+        if (!Arrays.stream(fieldsTask.split(",")).allMatch(field -> ShowTask.ALL_ATTRIBUTES.toLowerCase().contains(field.toLowerCase())))
+            throw new BadRequestException("The tasks' fields are invalid.");
         return new ShowGroup(group, groupService.getShowUserFromGroup(group)).getFields(fieldsGroup, fieldsUser, fieldsTask);
     }
 
     @PostMapping
     public Map<String, Object> addGroup(@RequestBody @Valid Group group) {
-        Preconditions.checkArgument(group.getName() != null && !Objects.equals(group.getName(), ""), "The group with idGroup " + group.getIdGroup() + " must have name.");
+        if (group.getName() == null || group.getName().isEmpty())
+            throw new BadRequestException("The group with idGroup " + group.getIdGroup() + " must have name.");
         if (group.getCreatedDate() == null) group.setCreatedDate(LocalDate.now().format(DateTimeFormatter.ISO_DATE));
         group = groupService.saveGroup(group);
         return new ShowGroup(group, groupService.getShowUserFromGroup(group)).getFields(ShowGroup.ALL_ATTRIBUTES, ShowUser.ALL_ATTRIBUTES, ShowTask.ALL_ATTRIBUTES);
@@ -96,7 +107,7 @@ public class GroupController {
     @PutMapping
     public Map<String, Object> updateGroup(@RequestBody @Valid Group group) {
         Group oldGroup = groupService.findGroupById(group.getIdGroup());
-        Preconditions.checkNotNull(oldGroup, "The group with idGroup " + group.getIdGroup() + " does not exist.");
+        if (oldGroup == null) throw new BadRequestException("The group with idGroup " + group.getIdGroup() + " does not exist.");
         if (group.getName() != null && !Objects.equals(group.getName(), ""))
             oldGroup.setName(group.getName());
         if (group.getDescription() != null && !Objects.equals(group.getDescription(), ""))
@@ -113,7 +124,7 @@ public class GroupController {
     @DeleteMapping("/{idGroup}")
     public Map<String, Object> deleteGroup(@PathVariable("idGroup") @Min(value = 0, message = "The idGroup must be positive.") Long idGroup) {
         Group group = groupService.findGroupById(idGroup);
-        Preconditions.checkNotNull(group, "The group with idGroup " + idGroup + " does not exist.");
+        if (group == null) throw new BadRequestException("The group with idGroup " + idGroup + " does not exist.");
         groupService.deleteGroup(group);
         return new ShowGroup(group, groupService.getShowUserFromGroup(group)).getFields(ShowGroup.ALL_ATTRIBUTES, ShowUser.ALL_ATTRIBUTES, ShowTask.ALL_ATTRIBUTES);
     }
@@ -123,9 +134,9 @@ public class GroupController {
     public Map<String, Object> addUserFromGroup(@PathVariable("idGroup") @Min(value = 0, message = "The idGroup must be positive.") Long idGroup,
                                                 @PathVariable("idUser") @Min(value = 0, message = "The idUser must be positive.") Long idUser) {
         Group group = groupService.findGroupById(idGroup);
-        Preconditions.checkNotNull(group, "The group with idGroup " + idGroup + " does not exist.");
+        if (group == null) throw new BadRequestException("The group with idGroup " + idGroup + " does not exist.");
         User user = userService.findUserById(idUser);
-        Preconditions.checkNotNull(user, "The user with idUser " + idUser + " does not exist.");
+        if (user == null) throw new BadRequestException("The user with idUser " + idUser + " does not exist.");
         groupService.addUserToGroup(group, user);
         return new ShowGroup(group, groupService.getShowUserFromGroup(group)).getFields(ShowGroup.ALL_ATTRIBUTES, ShowUser.ALL_ATTRIBUTES, ShowTask.ALL_ATTRIBUTES);
     }
@@ -135,9 +146,9 @@ public class GroupController {
     public Map<String, Object> deleteUserFromGroup(@PathVariable("idGroup") @Min(value = 0, message = "The idGroup must be positive.") Long idGroup,
                                                    @PathVariable("idUser") @Min(value = 0, message = "The idUser must be positive.") Long idUser) {
         Group group = groupService.findGroupById(idGroup);
-        Preconditions.checkNotNull(group, "The group with idGroup " + idGroup + " does not exist.");
+        if (group == null) throw new BadRequestException("The group with idGroup " + idGroup + " does not exist.");
         User user = userService.findUserById(idUser);
-        Preconditions.checkNotNull(user, "The user with idUser " + idUser + " does not exist.");
+        if (user == null) throw new BadRequestException("The user with idUser " + idUser + " does not exist.");
         groupService.removeUserFromGroup(group, user);
         return new ShowGroup(group, groupService.getShowUserFromGroup(group)).getFields(ShowGroup.ALL_ATTRIBUTES, ShowUser.ALL_ATTRIBUTES, ShowTask.ALL_ATTRIBUTES);
     }
@@ -147,9 +158,9 @@ public class GroupController {
     public Map<String, Object> addTaskFromGroup(@PathVariable("idGroup") @Min(value = 0, message = "The idGroup must be positive.") Long idGroup,
                                                 @PathVariable("idTask") @Min(value = 0, message = "The idTask must be positive.") Long idTask) {
         Group group = groupService.findGroupById(idGroup);
-        Preconditions.checkNotNull(group, "The group with idGroup " + idGroup + " does not exist.");
+        if (group == null) throw new BadRequestException("The group with idGroup " + idGroup + " does not exist.");
         Task task = taskService.findTaskById(idTask);
-        Preconditions.checkNotNull(task, "The task with idTask " + idTask + " does not exist.");
+        if (task == null) throw new BadRequestException("The task with idTask " + idTask + " does not exist.");
         groupService.addTaskToGroup(group, task);
         return new ShowGroup(group, groupService.getShowUserFromGroup(group)).getFields(ShowGroup.ALL_ATTRIBUTES, ShowUser.ALL_ATTRIBUTES, ShowTask.ALL_ATTRIBUTES);
     }
@@ -159,9 +170,9 @@ public class GroupController {
     public Map<String, Object> deleteTaskFromGroup(@PathVariable("idGroup") @Min(value = 0, message = "The idGroup must be positive.") Long idGroup,
                                                    @PathVariable("idTask") @Min(value = 0, message = "The idTask must be positive.") Long idTask) {
         Group group = groupService.findGroupById(idGroup);
-        Preconditions.checkNotNull(group, "The group with idGroup " + idGroup + " does not exist.");
+        if (group == null) throw new BadRequestException("The group with idGroup " + idGroup + " does not exist.");
         Task task = taskService.findTaskById(idTask);
-        Preconditions.checkNotNull(task, "The task with idTask " + idTask + " does not exist.");
+        if (task == null) throw new BadRequestException("The task with idTask " + idTask + " does not exist.");
         groupService.removeTaskFromGroup(group, task);
         return new ShowGroup(group, groupService.getShowUserFromGroup(group)).getFields(ShowGroup.ALL_ATTRIBUTES, ShowUser.ALL_ATTRIBUTES, ShowTask.ALL_ATTRIBUTES);
     }
@@ -170,7 +181,7 @@ public class GroupController {
     @DeleteMapping("/{idGroup}/users")
     public Map<String, Object> deleteAllUsersFromGroup(@PathVariable("idGroup") @Min(value = 0, message = "The idGroup must be positive.") Long idGroup) {
         Group group = groupService.findGroupById(idGroup);
-        Preconditions.checkNotNull(group, "The group with idGroup " + idGroup + " does not exist.");
+        if (group == null) throw new BadRequestException("The group with idGroup " + idGroup + " does not exist.");
         groupService.removeAllUsersFromGroup(group);
         return new ShowGroup(group, groupService.getShowUserFromGroup(group)).getFields(ShowGroup.ALL_ATTRIBUTES, ShowUser.ALL_ATTRIBUTES, ShowTask.ALL_ATTRIBUTES);
     }
@@ -179,9 +190,10 @@ public class GroupController {
     @GetMapping("/user/{idUser}")
     public List<Map<String, Object>> getGroupsWithUser(@PathVariable("idUser") @Min(value = 0, message = "The idUser must be positive.") Long idUser) {
         User user = userService.findUserById(idUser);
-        Preconditions.checkNotNull(user, "The user with idUser " + idUser + " does not exist.");
+        if (user == null) throw new BadRequestException("The user with idUser " + idUser + " does not exist.");
         List<Group> groups = groupService.findGroupsWithUser(user);
-        Preconditions.checkNotNull(groups, "The user with idUser " + idUser + " does not belong to any group.");
+        if (groups == null || groups.isEmpty())
+            throw new BadRequestException("The user with idUser " + idUser + " does not belong to any group.");
         return groups.stream().map(group -> new ShowGroup(group, groupService.getShowUserFromGroup(group)).getFields(ShowGroup.ALL_ATTRIBUTES, ShowUser.ALL_ATTRIBUTES, ShowTask.ALL_ATTRIBUTES)).toList();
     }
 
@@ -189,9 +201,10 @@ public class GroupController {
     @GetMapping("/task/{idTask}")
     public List<Map<String, Object>> getGroupsWithTask(@PathVariable("idTask") @Min(value = 0, message = "The idTask must be positive.") Long idTask) {
         Task task = taskService.findTaskById(idTask);
-        Preconditions.checkNotNull(task, "The task with idTask " + idTask + " does not exist.");
+        if (task == null) throw new BadRequestException("The task with idTask " + idTask + " does not exist.");
         List<Group> groups = groupService.findGroupsWithTask(task);
-        Preconditions.checkNotNull(groups, "The task with idTask " + idTask + " does not belong to any group.");
+        if (groups == null || groups.isEmpty())
+            throw new BadRequestException("The task with idTask " + idTask + " does not belong to any group.");
         return groups.stream().map(group -> new ShowGroup(group, groupService.getShowUserFromGroup(group)).getFields(ShowGroup.ALL_ATTRIBUTES, ShowUser.ALL_ATTRIBUTES, ShowTask.ALL_ATTRIBUTES)).toList();
     }
 }
